# fix(asm): fix reading wsgi input