--- a/ddtrace/contrib/flask/patch.py
+++ b/ddtrace/contrib/flask/patch.py
@@ -1,6 +1,7 @@
 import json
 
 import flask
+from six import BytesIO
 import werkzeug
 from werkzeug.exceptions import BadRequest
 import xmltodict
@@ -128,6 +129,13 @@ class _FlaskWSGIMiddleware(_DDWSGIMiddlewareBase):
         req_body = None
         if config._appsec_enabled and request.method in _BODY_METHODS:
             content_type = request.content_type
+            wsgi_input = environ.get("wsgi.input", "")
+
+            # Copy wsgi input if not seekable
+            if not wsgi_input.seekable():
+                body = wsgi_input.read()
+                environ["wsgi.input"] = BytesIO(body)
+
             try:
                 if content_type == "application/json":
                     if _HAS_JSON_MIXIN and hasattr(request, "json"):
@@ -144,15 +152,14 @@ class _FlaskWSGIMiddleware(_DDWSGIMiddlewareBase):
                     req_body = request.form.to_dict()
                 else:
                     req_body = request.get_data()
-                # Reset wsgi input to the beginning
-                wsgi_input = environ.get("wsgi.input")
-                if wsgi_input:
-                    try:
-                        wsgi_input.seek(0)
-                    except OSError:
-                        pass
             except (AttributeError, RuntimeError, TypeError, BadRequest):
                 log.warning("Failed to parse werkzeug request body", exc_info=True)
+            finally:
+                # Reset wsgi input to the beginning
+                if wsgi_input.seekable():
+                    wsgi_input.seek(0)
+                else:
+                    environ["wsgi.input"] = BytesIO(body)
 
         trace_utils.set_http_meta(
             span,