--- a/releasenotes/notes/asm-fix-reset-wsgi-input-035e0a7d917af2b2.yaml
+++ b/releasenotes/notes/asm-fix-reset-wsgi-input-035e0a7d917af2b2.yaml
@@ -1,0 +1,4 @@
+---
+fixes:
+  - |
+    Reset wsgi input after reading.