--- a/tests/contrib/flask/test_flask_appsec.py
+++ b/tests/contrib/flask/test_flask_appsec.py
@@ -1,5 +1,7 @@
 import json
 
+from flask import request
+
 from ddtrace.ext import http
 from ddtrace.internal import _context
 from ddtrace.internal.compat import urlencode
@@ -130,12 +132,21 @@ class FlaskAppSecTestCase(BaseFlaskTestCase):
         assert root_span.get_tag(http.USER_AGENT) == "test/1.2.3"
 
     def test_flask_body_urlencoded(self):
+        @self.app.route("/body", methods=["GET", "POST", "DELETE"])
+        def body():
+            data = request.form
+            return data, 200
+
         with override_global_config(dict(_appsec_enabled=True)):
             self.tracer._appsec_enabled = True
             # Hack: need to pass an argument to configure so that the processors are recreated
             self.tracer.configure(api_version="v0.4")
             payload = urlencode({"mytestingbody_key": "mytestingbody_value"})
-            self.client.post("/", data=payload, content_type="application/x-www-form-urlencoded")
+
+            response = self.client.post("/body", data=payload, content_type="application/x-www-form-urlencoded")
+            assert response.status_code == 200
+            assert response.json == payload
+
             root_span = self.pop_spans()[0]
             query = dict(_context.get_item("http.request.body", span=root_span))
 
@@ -165,12 +176,21 @@ class FlaskAppSecTestCase(BaseFlaskTestCase):
             assert query == {"attack": "1' or '1' = '1'"}
 
     def test_flask_body_json(self):
+        @self.app.route("/body", methods=["GET", "POST", "DELETE"])
+        def body():
+            data = request.get_json()
+            return data, 200
+
         with override_global_config(dict(_appsec_enabled=True)):
             self.tracer._appsec_enabled = True
             # Hack: need to pass an argument to configure so that the processors are recreated
             self.tracer.configure(api_version="v0.4")
             payload = {"mytestingbody_key": "mytestingbody_value"}
-            self.client.post("/", json=payload, content_type="application/json")
+
+            response = self.client.post("/body", json=payload, content_type="application/json")
+            assert response.status_code == 200
+            assert response.json == payload
+
             root_span = self.pop_spans()[0]
             query = dict(_context.get_item("http.request.body", span=root_span))
 
@@ -190,12 +210,20 @@ class FlaskAppSecTestCase(BaseFlaskTestCase):
             assert query == {"attack": "1' or '1' = '1'"}
 
     def test_flask_body_xml(self):
+        @self.app.route("/body", methods=["GET", "POST", "DELETE"])
+        def body():
+            data = request.data
+            return data, 200
+
         with override_global_config(dict(_appsec_enabled=True)):
             self.tracer._appsec_enabled = True
             # Hack: need to pass an argument to configure so that the processors are recreated
             self.tracer.configure(api_version="v0.4")
             payload = "<mytestingbody_key>mytestingbody_value</mytestingbody_key>"
-            self.client.post("/", data=payload, content_type="application/xml")
+            response = self.client.post("/body", data=payload, content_type="application/xml")
+            assert response.status_code == 200
+            assert response.data == b"<mytestingbody_key>mytestingbody_value</mytestingbody_key>"
+
             root_span = self.pop_spans()[0]
             query = dict(_context.get_item("http.request.body", span=root_span))