@avara1986 2022-08-18
Benchmarks of appsec enabled:

```
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| Benchmark                       | /artifacts/41d46418-2f38-4a52-91d5-819bfbad8552/flask_simple/1.4.0rc2//results.json | /artifacts/41d46418-2f38-4a52-91d5-819bfbad8552/flask_simple/1.4.0rc2.dev60+g4b147835//results.json |
+=================================+=====================================================================================+=====================================================================================================+
| flasksimple-baseline            | 1.97 ms                                                                             | 2.23 ms: 1.13x slower                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| flasksimple-tracer              | 2.34 ms                                                                             | 1.91 ms: 1.23x faster                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| flasksimple-profiler            | 2.55 ms                                                                             | 1.91 ms: 1.34x faster                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| flasksimple-appsec-get          | 2.24 ms                                                                             | 2.47 ms: 1.10x slower                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| flasksimple-appsec-post         | 2.36 ms                                                                             | 2.16 ms: 1.09x faster                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| flasksimple-tracer-and-profiler | 2.24 ms                                                                             | 2.32 ms: 1.04x slower                                                                               |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+
| Geometric mean                  | (ref)                                                                               | 1.04x faster                                                                                        |
+---------------------------------+-------------------------------------------------------------------------------------+-----------------------------------------------------------------------------------------------------+

Benchmark hidden because not significant (1): flasksimple-tracer-and-profiler-and-appsec
```