@avara1986 2022-08-18
> flask is the only server which is having this problem?

Yes, Django and Pylons have other approach when you try to retrieve the body