@avara1986 2022-08-22
> Can we run the new benchmarks and include the results in the PR description? It seems like an interesting to document

why are these result not valid? https://github.com/DataDog/dd-trace-py/pull/4114#issuecomment-1219249212